# Optimize Ordinal.EqualsIgnoreCase_Vector with AVX2 and AVX512