--- a/src/libraries/System.Private.CoreLib/src/System/Globalization/Ordinal.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/Globalization/Ordinal.cs
@@ -82,7 +82,6 @@ namespace System.Globalization
             where TVector : struct, ISimdVector<TVector, ushort>
         {
             Debug.Assert(length >= TVector.Count);
-            Debug.Assert(TVector.IsHardwareAccelerated);
 
             nuint lengthU = (nuint)length;
             nuint lengthToExamine = lengthU - (nuint)TVector.Count;