Closes https://github.com/dotnet/runtime/issues/92497

Currently, `str1.Equals(str2, StringComparison.OrdinalIgnoreCase)` is only accelerated with `Vector128`, this PR adds `Vector256` and `Vector512` via `ISimdVector` abstraction - the best part that there is no code duplication.


```cs
public class Benchmarks
{
    static readonly string s1 = @"C:\prj\runtime\src\coreclr\System.Private.CoreLib\System.Private.CoreLib.csproj";
    static readonly string s2 = @"C:\prj\runtime\src\coreclr\System.Private.CoreLib\system.Private.CoreLib.csproj";

    [Benchmark]
    public bool EqualsIgnoreCase()
    {
        return s1.Equals(s2, StringComparison.OrdinalIgnoreCase);
    }
}
```

| Method           | Toolchain             | Mean      | Error    | StdDev    | Ratio |
|----------------- |---------------------- |----------:|---------:|----------:|------:|
| EqualsIgnoreCase | Core_Root\corerun.exe | 10.610 ns | 0.1701 ns | 0.1591 ns |  1.00 |
| EqualsIgnoreCase | Core_Root_PR\corerun.exe |  **5.878 ns** | 0.0354 ns | 0.0313 ns |  0.55 |