@EgorBo 2023-10-07
@stephentoub @tannergooding PTAL, some notes:
1) I copied parts of `Ascii.EqualsIgnoreCase` here, I can't just invoke `Ascii.EqualsIgnoreCase` because we need to fallback to ICU in case of non-ascii, while `Ascii.EqualsIgnoreCase` just bails out. I'll think of how I can unify them separately.
2) Trailing elements are now handled with SIMD (used to be scalar) so all benchmarks for Length=8 (chars) and higher show improvements.

Had to replace == with Equals due to https://github.com/dotnet/runtime/issues/93174