@EgorBo 2023-10-12
Ping, also cc @adamsitnik @MihaZupan who authored/contributed Ascii.cs where I copied this code from